separate "Create Task" and "View Task" functionality
