+++ app/src/main/java/com/example/FocusApp/FocusDestinations.kt
@@ -14,11 +14,11 @@ object Stats: FocusDestination {
 }
 
 object Tasks: FocusDestination {
-    override val route = "tasks"
+    override val route = "Add Task"
 }
 
 object Generators: FocusDestination {
-    override val route = "generator"
+    override val route = "My Tasks"
 
 }
 object Login: FocusDestination {
@@ -28,4 +28,4 @@ object Login: FocusDestination {
 
 
 // Screens to be displayed in the top RallyTabRow
-val focusTabRowScreens = listOf(Tasks, Stats, Generators, Login)
+val focusTabRowScreens = listOf(Generators,Stats, Login)
